Rename MetricsFileParser
